% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{costs_to_display}
\alias{costs_to_display}
\alias{get_nested_costs}
\alias{get_life_years_lost}
\alias{cost_item}
\title{Prepare Daedalus costs output for display}
\usage{
get_nested_costs(raw_costs)

get_life_years_lost(raw_costs)

cost_item(id, values, children = NULL)
}
\arguments{
\item{raw_costs}{A list resulting from a call to \code{daedalus::get_costs()} on
a \verb{<daedalus_output>} class object.}

\item{id}{String description of list name.}

\item{values}{List contents.}

\item{children}{Nested lists contained within the top-level list, if any.
Defaults to \code{NULL}.}
}
\value{
A recursive nested list with the elements \code{"id"} and \code{"value"}
with a string and numeric value respectively.
Additionally, a \code{"children"} list element may be present containing another
list with the same recursive structure.
The lowest level list within \code{"children"} has no \code{"children"} element.
}
\description{
\code{get_nested_costs()} and \code{get_nested_natural_costs()} prepare daedalus costs
outputs for display. \code{get_natural_costs()} currently only prepares life-years
lost but may include other costs in their natural units in future.

\code{cost_item()} is a helper function that prepares list elements in the format
\code{"id"}, \code{"value"}, \code{"children"}.
}
\keyword{internal}
