% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{custom_tests}
\alias{custom_tests}
\alias{expect_nested_value_sum}
\title{Expect values for nested costs output}
\usage{
expect_nested_value_sum(x, i_unit = 1L)
}
\arguments{
\item{x}{A nested list of costs, which is expected to have a top level cost
under \code{x$values[[i]]$value}, and nested costs under \code{x$children}.}

\item{i_unit}{An index, defaulting to 1, for which index of \code{x$values} is
summed.}
}
\description{
Helper expectations functions to check nested costs output.
\code{expect_nested_value_sum()} checks that the level cost under
\code{x$values[[1]]$value} is the sum of the costs associated with its children.

\code{expect_nested_names()} checks that \code{x$children} have expected names.
}
\keyword{internal}
