% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.api-package.R
\docType{package}
\name{daedalus.api-package}
\alias{daedalus.api}
\alias{daedalus.api-package}
\title{daedalus.api: Serve the 'daedalus' model via an API}
\description{
API server for the 'daedalus' model, using 'porcelain' and 'plumber'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameel-institute/daedalus.api}
  \item Report bugs at \url{https://github.com/jameel-institute/daedalus.api/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{p.gupte24@imperial.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

Authors:
\itemize{
  \item Emma Russell
  \item Rich FitzJohn (\href{https://orcid.org/0000-0001-8888-3837}{ORCID})
  \item David Mears
}

Other contributors:
\itemize{
  \item Abdul Latif Jameel Institute for Disease and Emergency Analytics [funder]
  \item Imperial College of Science, Technology and Medicine [copyright holder, funder]
}

}
\keyword{internal}
