% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\title{Class for managing running jobs on the redis queue}
\description{
Class for managing running jobs on the redis queue

Class for managing running jobs on the redis queue
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{controller}}{RRQ controller}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Queue-new}{\code{Queue$new()}}
\item \href{#method-Queue-queue_model_run}{\code{Queue$queue_model_run()}}
\item \href{#method-Queue-get_run_status}{\code{Queue$get_run_status()}}
\item \href{#method-Queue-get_run_results}{\code{Queue$get_run_results()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-new"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-new}{}}}
\subsection{Method \code{new()}}{
Initialise redis connection and rrq.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$new(queue_id = NULL, separate_process = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-queue_model_run"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-queue_model_run}{}}}
\subsection{Method \code{queue_model_run()}}{
Submit a model run job to the queue, and return the run id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$queue_model_run(parameters, model_version = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{parameter values for the model run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-get_run_status"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-get_run_status}{}}}
\subsection{Method \code{get_run_status()}}{
Get status information for a model run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$get_run_status(run_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run_id}}{the run id of the model run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-get_run_results"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-get_run_results}{}}}
\subsection{Method \code{get_run_results()}}{
Get results data for a completed model run. Throws an error if the task
was not successful.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$get_run_results(run_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run_id}}{the run id of the model run}
}
\if{html}{\out{</div>}}
}
}
}
