% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validate_country}
\alias{validate_country}
\title{Validate Country}
\usage{
validate_country(country)
}
\arguments{
\item{country}{A character string representing
the name of the country to validate.}
}
\value{
This function does not return a value.
It will stop execution if the validation fails.
}
\description{
This function checks if the provided country name is
valid by ensuring it is a character string
and exists within the \code{daedalus::country_names} list.
}
\examples{
\dontrun{
validate_country("United States")
validate_country("InvalidCountry") # This will cause an error
}
}
\keyword{internal}
