% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{RESULT_NAMES}
\alias{TIMESERIES_NAMES}
\title{Output constants}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
RESULT_NAMES

TIMESERIES_NAMES
}
\description{
Output constants
}
\keyword{constants}
