% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_annual_gdp}
\alias{get_annual_gdp}
\title{Get annual GDP from DAEDALUS country data}
\usage{
get_annual_gdp(country)
}
\arguments{
\item{country}{A string giving a country name
from among \code{daedalus::country_names}.}
}
\value{
A single number value for the annual GDP of a country in terms of
million dollars. Values are in 2018 terms.
}
\description{
Convert daily GVA values to annual GDP values.
}
\keyword{internal}
