% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_average_vsl}
\alias{get_average_vsl}
\title{Get Average Value of Statistical Life (VSL) for a Country}
\usage{
get_average_vsl(country)
}
\arguments{
\item{country}{A character string representing the name of the country.}
}
\value{
A numeric value representing the average
VSL for the specified country.
}
\description{
This function calculates the average
Value of Statistical Life (VSL) for a specified country.
It computes the weighted mean of VSL using the demography data as weights.
}
\examples{
\dontrun{
  avg_vsl <- get_average_vsl("USA")
  print(avg_vsl)
}

}
\keyword{internal}
