% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour.R
\name{process_behaviour_choice}
\alias{process_behaviour_choice}
\title{Translate behaviour choice to daedalus input}
\usage{
process_behaviour_choice(x, hospital_capacity)
}
\arguments{
\item{x}{A single string for the behavioural choice.}

\item{hospital_capacity}{The country hospital capacity.}
}
\value{
Either \code{NULL}, or a \verb{<daedalus_behaviour>} with 'new' behavioural
parameters, suitable for passing to the \code{behaviour} argument of
\code{\link[daedalus:daedalus]{daedalus::daedalus()}}.
}
\description{
Translate behaviour choice to daedalus input
}
